Changed personnel search to a GET personnel route
