+++ server/api/units/index.js
@@ -25,14 +25,6 @@ router.get(
 );
 
 router.get(
-  '/search',
-  auth.isApiAuthenticated,
-  auth.hasRole('user'),
-  auth.hasFireDepartment,
-  metricsController.search
-);
-
-router.get(
   '/:id/metrics',
   auth.isApiAuthenticated,
   auth.hasRole('user'),
