+++ server/api/units/unit-metrics.controller.js
@@ -276,26 +276,3 @@ export function rankBuckets(buckets) {
 
   return rankedBuckets;
 }
-
-export async function search(req, res) {
-  const { query } = req.query;
-  
-  if (!query) {
-    return res.json([]);
-  }
-
-  const index = req.fireDepartment.es_indices['fire-incident'];
-  const searchBody = bodybuilder()
-    .size(0)
-    .aggregation('terms', 'description.resources.personnel.dispatched', { size: 10000 })
-    .build();
-
-  const esRes = await connection.getClient().search({
-    index,
-    body: searchBody,
-  });
-
-  const buckets = _.get(esRes, 'aggregations["agg_terms_description.resources.personnel.dispatched"].buckets');
-  const personel = buckets.map(bucket => bucket.key).filter(key => key.includes(query));
-  return res.json(personel);
-};
