+++ server/routes.js
@@ -28,6 +28,7 @@ export default function(app) {
   app.use('/api/fire-departments', require('./api/fire-department'));
   app.use('/api/incidents', require('./api/incident'));
   app.use('/api/units', require('./api/units'));
+  app.use('/api/personnel', require('./api/personnel'));
   app.use('/api/twitter', require('./api/twitter'));
   app.use('/api/users', require('./api/user'));
   app.use('/api/reports', require('./api/report'));
